refactored the engine of the default models
I moved the attention mechanism and the engine for the ResNet into separate files.

The example config file and the benchmark config files are updated accordingly.
