+++ config/v_0_5_0_benchmarks/particle_nocut.yml
@@ -33,11 +33,11 @@ Model:
         particletype: {class_names: ['proton', 'gamma'], fc_head: [512, 256, 2], weight: 1.0}
 Model Parameters:
     res_net:
-        network: {module: 'basic', function: 'stacked_res_blocks'}
+        network: {module: 'resnet_engine', function: 'stacked_res_blocks'}
         #init_layer: {filters: 64, kernel_size: 7, strides: 1}
         #init_max_pool: {size: 3, strides: 2}
         pretrained_weights: null
-    single_tel:
+    single_cnn:
         network: {module: 'basic', function: 'conv_block'}
         pretrained_weights: null
     cnn_rnn:
@@ -69,14 +69,16 @@ Model Parameters:
                 - {filters: 256, kernel_size: 3}
             final_avg_pool: true
             batchnorm: false
+        batchnorm_decay: 0.99
+    resnet_engine:
         stacked_res_blocks:
+            residual_block: 'bottleneck'
             architecture:
                 - {filters: 48, blocks: 2}
                 - {filters: 96, blocks: 3}
                 - {filters: 128, blocks: 3}
                 - {filters: 256, blocks: 3}
-        attention: {mechanism: 'Squeeze-and-Excitation', ratio: 16}
-        batchnorm_decay: 0.99
+    attention: {mechanism: 'Squeeze-and-Excitation', ratio: 16}
 Training:
     validation_split: 0.05
     num_epochs: 6
